[codex] Refactor progress-sync cleanup
